From 5ca31a1f4ad9dfa4490b8ea264ca42903a98f004 Mon Sep 17 00:00:00 2001 Subject: [PATCH 39/54] powerpc/fsl: Use new clockgen binding
The driver retains compatibility with old device trees, but we don't
want the old nodes lying around to be copied, or used as a reference
(some of the mux options are incorrect), or even just being clutter.

We will also need the #clock-cells in the clockgen node in order to
add fman nodes.

Signed-off-by: Scott Wood <scottwood@freescale.com>